* replace int with size_t in places where MSVS flags a warning [rtj]

+++ src/xml-hddm.cpp
@@ -356,8 +356,8 @@ void HDDMmaker::constructDocument(DOMElement* el)
    XString tagS(el->getTagName());
    *ofs << "<" << tagS;
    DOMNamedNodeMap* attrList = el->getAttributes();
-   int attrListLength = attrList->getLength();
-   for (int a = 0; a < attrListLength; a++)
+   size_t attrListLength = attrList->getLength();
+   for (size_t a = 0; a < attrListLength; a++)
    {
       DOMNode* node = attrList->item(a);
       XString nameS(node->getNodeName());
@@ -366,12 +366,12 @@ void HDDMmaker::constructDocument(DOMElement* el)
    }
 
    DOMNodeList* contList = el->getChildNodes();
-   int contListLength = contList->getLength();
+   size_t contListLength = contList->getLength();
    if (contListLength > 0)
    {
       *ofs << ">" << std::endl;
       indent++;
-      for (int c = 0; c < contListLength; c++)
+      for (size_t c = 0; c < contListLength; c++)
       {
          DOMNode* node = contList->item(c);
          if (node->getNodeType() == DOMNode::ELEMENT_NODE)
@@ -402,12 +402,12 @@ void HDDMmaker::outputStream(DOMElement* thisEl, DOMElement* modelEl,
    XString thisS(thisEl->getTagName());
 
    DOMNamedNodeMap* modelAttrList = modelEl->getAttributes();
-   int modelAttrListLength = modelAttrList->getLength();
+   size_t modelAttrListLength = modelAttrList->getLength();
    DOMNamedNodeMap* thisAttrList = thisEl->getAttributes();
-   int thisAttrListLength = thisAttrList->getLength();
+   size_t thisAttrListLength = thisAttrList->getLength();
    XString minS(modelEl->getAttribute(X("minOccurs")));
    XString maxS(modelEl->getAttribute(X("maxOccurs")));
-   int expectAttrCount = modelAttrList->getLength()
+   size_t expectAttrCount = modelAttrList->getLength()
                          - (minS == ""? 0 : 1)
                          - (maxS == ""? 0 : 1);
    if (thisAttrListLength != expectAttrCount)
@@ -521,10 +521,10 @@ void HDDMmaker::outputStream(DOMElement* thisEl, DOMElement* modelEl,
    }
 
    DOMNodeList* thisList = thisEl->getChildNodes();
-   int thisListLength = thisList->getLength();
+   size_t thisListLength = thisList->getLength();
    DOMNodeList* modelList = modelEl->getChildNodes();
-   int modelListLength = modelList->getLength();
-   for (int m = 0; m < modelListLength; m++)
+   size_t modelListLength = modelList->getLength();
+   for (size_t m = 0; m < modelListLength; m++)
    {
       DOMNode* mode = modelList->item(m);
       short type = mode->getNodeType();
